+++ internal/provider/model/job_definitions/filter/filter_mask.go
