+++ internal/provider/model/job_definitions/input_options/decoder.go
