+++ internal/provider/model/job_definitions/input_options/parser/jsonpath_parser.go
