+++ internal/provider/model/job_definitions/output_options/bigquery.go
